Incorporate internal function db_pivot_wider for the time being *
It appears that another pivot_wider function is being prepared at dbplyr,
but in the meantime I wanted to prevent the need for my fork of dbplyr, that
took the pivot_wider branch of Edgar Ruiz (who proposed the function)
and brought it up to date with dbplyr's master at times.
By incorporating the function, the watina package can just work with
dbplyr from CRAN and still use the functionality.
All credits for db_pivot_wider() go to Edgar Ruiz.